Added: Lombok jar file is added to classpath to generate getter and setter in ResponseDTO.java
