+++ EmployeePayrollApp/src/main/java/com/cg/employeepayrollapp/dto/ResponseDTO.java
@@ -1,0 +1,13 @@
+package com.cg.employeepayrollapp.dto;
+
+import lombok.Data;
+
+public @Data class ResponseDTO {
+	private String message;
+	private Object data;
+
+	public ResponseDTO(String message, Object data) {
+		this.message = message;
+		this.data = data;
+	}
+}
