--- EmployeePayrollApp/src/main/java/com/cg/employeepayrollapp/model/ResponseDTO.java
@@ -1,28 +1,0 @@
-package com.cg.employeepayrollapp.model;
-
-public class ResponseDTO {
-	private String message;
-	private Object data;
-
-	public ResponseDTO(String message, Object data) {
-		super();
-		this.message = message;
-		this.data = data;
-	}
-
-	public String getMessage() {
-		return message;
-	}
-
-	public void setMessage(String message) {
-		this.message = message;
-	}
-
-	public Object getData() {
-		return data;
-	}
-
-	public void setData(Object data) {
-		this.data = data;
-	}
-}
